+++ tests/posix/open_memstream.c
@@ -11,14 +11,14 @@ int main() {
 	assert(fp);
 
 	char c = 'A';
-	for (int i = 0; i < WRITE_NO; i++)
+	for (size_t i = 0; i < WRITE_NO; i++)
 		assert(fwrite(&c, sizeof(char), 1, fp) == 1);
 
 	// Flush the file to update the pointers.
 	assert(!fflush(fp));
 
 	assert(size == WRITE_NO);
-	for (int i = 0; i < size; i++)
+	for (size_t i = 0; i < size; i++)
 		assert(buf[i] == c);
 
 	// Check if the stream maintains a null-bye at the end.
@@ -31,29 +31,29 @@ int main() {
 	assert(buf[size] == '\0');
 
 	// Check if it's filled with zero's.
-	for (int i = WRITE_NO; i < size; i++)
+	for (size_t i = WRITE_NO; i < size; i++)
 		assert(buf[i] == '\0');
 
 	// Go back and overwrite the 0's with 'B'.
 	assert(!fseek(fp, -WRITE_NO, SEEK_CUR));
 	c = 'B';
-	for (int i = 0; i < WRITE_NO; i++)
+	for (size_t i = 0; i < WRITE_NO; i++)
 		assert(fwrite(&c, sizeof(char), 1, fp) == 1);
 
 	// Check if that happened.
 	assert(size == 2*WRITE_NO);
-	for (int i = WRITE_NO; i < size; i++)
+	for (size_t i = WRITE_NO; i < size; i++)
 		assert(buf[i] == c);
 	assert(buf[size] == '\0');
 
 	// Go to the front and write 'B'.
 	assert(!fseek(fp, 0, SEEK_SET));
-	for (int i = 0; i < WRITE_NO; i++)
+	for (size_t i = 0; i < WRITE_NO; i++)
 		assert(fwrite(&c, sizeof(char), 1, fp) == 1);
 
 	// Check if that happened.
 	assert(size == 2*WRITE_NO);
-	for (int i = 0; i < size; i++)
+	for (size_t i = 0; i < size; i++)
 		assert(buf[i] == c);
 	assert(buf[size] == '\0');
